Add spotbugs support
A new plugin and feature for spotbugs is added, so both findbugs and
spotbugs will work. Spotbugs and findbugs implementation is kept totally
separated, so that future modifications for spotbugs will not affect
the deprecated findbugs support.

The next commit will introduce some refactoring to better reuse some
infrastructure between the plugins (and checktsyle and pmd, too)

Fixes #110