Fix #203: Have org.apache.commons.lang in targets
The POM repositories and the Eclipse targets are somewhat out of sync.
As a quickfix just add the missing bundle to the Eclipse target. In the
long run I suggest to use a .target file in the Maven build, too,
instead of fetching whatever is available on the repository URLs.